Удален RealmSwift
