+++ AuthorizationForm/App/AppDelegate.swift
@@ -1,0 +1,20 @@
+//
+//  AppDelegate.swift
+//  AuthorizationForm
+//
+//  Created by Александр Меренков on 4/6/22.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
+
