+++ AuthorizationForm/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -86,6 +86,7 @@
       "size" : "83.5x83.5"
     },
     {
+      "filename" : "1024.png",
       "idiom" : "ios-marketing",
       "scale" : "1x",
       "size" : "1024x1024"
