--- AuthorizationForm/Model/Token.swift
@@ -1,12 +1,0 @@
-//
-//  Token.swift
-//  AuthorizationForm
-//
-//  Created by Александр Меренков on 4/12/22.
-//
-
-import RealmSwift
-
-class Token: Object {
-    @objc dynamic var token = ""
-}
