--- AuthorizationForm/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  AuthorizationForm
-//
-//  Created by Александр Меренков on 4/6/22.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
